Let the fun begin
Added new (empty) classes to show how the game will eventually be
structured. Map generation is nearing completion, still need to
determine SOLIDS and update the map squares accordingly. Need to decide
what will be sent across the network. Eventually, and this is a long
ways out, the code will create maps and players based on resource files
that will likely be .XML files with attributes specifying how to
generate the objects.