Changed some image names
Next step is to make resource gathering dynamic. Preferably have a .xml
file with values that will act as parameters for the methods used to
create a map.
Also need to work on the format for images, add special groups which are
defined as a collection of sub-images that need to be organized in a
certain way to create a single, larger object.
Create an outside-level using the outdoor textures. (New code will have
to be added to allow for different types of terrain generation.)
Inanimate (SOLID) objects (tables, chairs, trees) need to have methods
implemented for generating and adding them.