Map Objects
I have added a lot of new (mostly unmanaged) texture resources.

I have also began implementing object generation and adding to the map.
Figuring out the best way to map the object to a mini-grid system using
each component of the object and then copying that into the actual map
is still the main issue needing to be solved.