A bit of cleanup
Cleaned up some files some.

IMPORTANT - This is Squares Interactive, not Squares. Need to begin
implementing the ability to "communicate" with other players. For now
this will likely include being in an adjacent square, facing them, and
hitting a key which will do some checks and then create a box where you
can type a message (escape to make the box go away) and hit enter. After
you hit enter, the message will appear above your avatar's head in a
bubble for everyone to see.

Also important, thought not crucial, is working on the sliding window
view which will center the player in the client window and add a "fog of
war" around the player, making them walk around the map to explore.

Also important, need to finalize object component texture mapping to a
logical grid and then dumping the grid into the map. We want some
obstacles!