Added a client-server communication
The GUI window waits for data to be received, and then moves the player
based on the data received.

When the user presses a key on the keyboard, a move request is sent to
the host. If the host says the move is valid, then it will broadcast the
move to all the clients which will grab the data in their receivers.